[Feat #267] 죽은 유저 이미지 캠 처리 및 리팩토링
