+++ types/index.ts
@@ -42,7 +42,7 @@ export interface VoteResult {
 
 export interface Participants {
   tracks: TrackReferenceOrPlaceholder[];
-  checkClickHandle: (event: React.MouseEvent<HTMLElement>, userId: string) => void;
+  // checkClickHandle: (event: React.MouseEvent<HTMLElement>, userId: string) => void;
 }
 
 export interface playerMedia {
